requirements.txt
streamlit==0.76.0
numpy==1.19.5
tensorflow_gpu==2.4.1
Pillow==8.1.0
tensorflow==2.4.1
